osd: error out current op only on tmap ENOENT
If an individual op in do_osd_ops() fails, set result, and continue to the
end of the loop.  Normally we'll return the error then, but if the ignore
error flag is set we'll continue.

Fix the tmap error handling so that we don't modify the tmap if the op
failed.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>