mds: fix possible deadlock in multi-mds setup
This should fix the file_excl case on a file_max update when there
is more than one mds.

If we dont issue caps here its possible when we eval_gather on
the returning lockack that we don't push the lock forward which
will leave the inode auth pinned. If an export occurs that tries
to freeze a tree this inode is in, we'll have a freezing tree
waiting for an auth_pin that can no longer be removed.

Signed-off-by: Brandon Seibel <brandon@seibelnet.ca>