rgw: don't specify create_pool and set_marker in create_bucket.
It's wildly inappropriate for that kind of implementation detail to
leak out of the interface. For the moment, leave in a "system_bucket"
parameter, which provides the necessary information. But we should
probably be able to push all of that information down below the
interface layer eventually, and then get rid of that param too.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>