osd: fix assemble_backlog
This was written assuming that le->prior_version wouldn't be the version
that we have locally on disk.  Not always true!

If it is the same, then we can just keep the entry (and clear reqid).  If
it is different, keep the behavior we had (re-add, erase current).

FWIW the last time this was touched was 916b1998.

Signed-off-by: Sage Weil <sage@newdream.net>