osd: fix add_next_event Missing::item::have
The missing set should be accurate up to the current point in the log.  The
log_tail has no bearing on that, nor does last_update, since we're
processing new events in forward order, and updating missing as we go.

Drops the now unused info argument... :/

This more or less reverts b418896d.

Signed-off-by: Sage Weil <sage@newdream.net>