cls_rgw: add constructors to data structs; don't leak tags on races
We were leaking tags on races before, since we cut out of the function
before clearing the tag. We don't do that any more.
Also, we now use a constructor for the dir entry structs so that
we don't follow the wrong paths (based on random memory) for
brand new objects.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>