osd: PgPriorSet: remove up_thru crap
This was added way back in 1cf9bebc8e5063f5f311d33e7735bcc9286e98ce,
but as far as I can tell it didn't make any sense then either.

The issue: we redo the prior set calculation if the up_thru for these
OSDs changes in the current map, but the prior_set result does not
depend on the current map's up_thru values in any way; it only depends
on the up_thru in the last epoch of each past interval, and that is
fixed in the past.

Remove the cruft.

Signed-off-by: Sage Weil <sage@newdream.net>