test_filestore_idempotent: simple tool to generate a worklaod of non-idempotent operations
Generate a workload of operations that are non-idempotent.  These are:

 transaction {
   clone A -> A.($n-1)
   write $n to A
 }
 $n++
 loop!

If we apply any transaction to the file system more than once, we will
find that the A.$n object does not contain $n, but instead contains
some larger value.

First run in 'write' mode to generate a workload and fake a crash.
Then run in 'verify' mode to see if the result was bad.

Signed-off-by: Sage Weil <sage@newdream.net>