osd: pg_pool_t: introduce flags, crash_replay_interval
Introduce a per-pool crash_replay_interval so we can control whether
the OSD waits for replayed ACKed but not COMMITted requests for this
PG.  For the metadata and rbd pools, for instance, the replay window
is useless.

Introduce a generic flags field, while we're modifying the encoding.

No new feature bit; piggyback on POOL3.

Signed-off-by: Sage Weil <sage@newdream.net>