osd: PgPriorSet: kill whoami; make PG arg strictly optional
It is only used for the debug output prefix.  Make it so we can leave it
out entirely (e.g. for unit tests).

We don't want to, say, pass in the string prefix itself, or else we are
stuck with generating that string even on low debug levels where it won't
be used.

Signed-off-by: Sage Weil <sage@newdream.net>