osd: PriorSet: acting/up membership implies still alive
If the osd is in the acting or up sets, we can assume they are still alive,
even though we don't know that for sure, because if they are not, we will
rebuild PriorSet.

Note that we have a dependency here on up_thru that we could/should rebuild
PriorSet based on, IF we think it might change the value of the CRASHED
flag and IF we care enough.  Right now we don't.  Marking CRASHED when we
don't need to is conservative, and not dangerous.

Signed-off-by: Sage Weil <sage@newdream.net>