osd: pg_pool_t: set crash_replay_interval on data pool when decoding old
We want to preserve the crash_replay_interval on old clusters being
upgraded.  Kludge this by setting it to 60 (the old default) if the
crush_ruleset == 0 and owner == 0, which is normally true for just the
data pool.

This may catch other pools they created by hand, but it's still better
than having the replay interval for all pools when it is not needed.

Signed-off-by: Sage Weil <sage@newdream.net>