uclient: fix _getdents and add some documentation.
If readdir_r_cb returns 0, that means SUCCESS, regardless of how
many entries it actually wrote.
If it returns <0, then either:
a) our callback failed to fit some data (returning -1), which is fine
unless we took no entries (so return gr.pos or -ERANGE, depending), or
b) something else failed, so return that error code.
If it returns 0, we succeeded, so return the amount of used buffer.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>