os: rename and make use of the split_threshold parameter.
This was accidentally left out of the must_split calculation. Put it
in, and rename it to split_multiplier (as that is a much better name
for how it's used).

In the default case this won't actually change behavior, but it
makes the behavior configurable as it's supposed to be.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>