doc: Import the list of ceph subcommands from wiki.
This adds the content of the wiki page at

http://ceph.newdream.net/wiki/Monitor_commands

to doc/control.rst in order to make it available at the new official
location for the ceph documentation. This first patch is just the
result of a cut-and-paste operation. There are no changes in content,
but the text was converted to rst format.

Signed-Off-By: Andre Noll <maan@systemlinux.org>