doc: Add missing documentation for osd pool get.
"osd pool set" was already documented, but the corresponding "get"
command was not. This patch adds the list of valid fields for this
command, together with short descriptions.

Thanks to Gregory Farnum for pointing out a flaw in the description
of lpl_num and lpgp_num in an earlier version of the patch.

Signed-Off-By: Andre Noll <maan@systemlinux.org>