osd: trigger RecoveryFinished event on recovery completion
Unconditionally trigger the RecoveryFinished event when start_recvoery_ops
thinks it may be done.  This lets us trigger the acting change (if needed),
or call finish_recovery() if needed.

This fixes the case where we are backfilling with up == acting, complete,
but don't call finish_recovery() or clear the backfill|degraded bits.

At some point we may want to move the is_all_uptodate() checks to the
caller.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>