mon: rev cluster protocol
The OSDMap NEW and AUTOOUT bit additions subtely change the decoding of
the incremental maps in a reasonably harmless way in that the bits get
implicitly cleared whenever the OSD weight changes from non-zero.  The
monitors need to agree on this behavior to avoid odd behavior.  We don't
care what clients see, since those bits are informational only.

Signed-off-by: Sage Weil <sage@newdream.net>