objecter: ignore replies from old request attempts
If we know the request attempt, ignore old attempts.

If we do not know the attempt (because the server is old), accept the
reply.  This could lead to doing some ACK callbacks we shouldn't in
extreme failure/recovery scenarios, but that is better than doing
the callbacks out of order.

Partially fixes: #1490
Signed-off-by: Sage Weil <sage@newdream.net>