ceph: add a new "run_uml.sh" script to manage running a UML client
This script is used to automate most of what's required to run a
User-Mode Linux (UML) instance.  This is mainly of interest for
ceph client developers who might benefit from the debugger access
that UML affords.  It was written for ceph development but isn't
really dependent on ceph.  It basically makes a few assumptions and
follows some conventions, and in doing so is able to encapsulate
most of the "tricky parts" of setting up to run a UML instance.

Signed-off-by: Alex Elder <elder@dreamhost.com>