mon: make PaxosService::update_from_paxos return void.
You can't really recover from a failed update (as PGMonitor was trying
to do), and nothing in the system checks the return values.
So rip out the return values and change failed updates to an assert
failure (most of the other Monitors continue to have decode exceptions,
but we can keep the pretty that we have).

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>