osd: introduce RECOVERING pg state
Since clean now means not degraded, we need some other indication that
recovery has completed and we are "done" (given the current up/down state
of the OSDs).

Adding a 'recovering' state also makes it clearer to users that work is
being done, as opposed to the current situation, where they look for the
absense of 'clean'.

Signed-off-by: Sage Weil <sage@newdream.net>