osd: make op_commit imply op_applied for purposes of repop completion
For repop completion, we want waitfor_ack and _commit to be empty.  For
replicas, a commit reply implies ack, so ack is always a subset of commit.
But for the local write, we wait for applied separately, so we can have
repops open where we sent the reply to the client but still have it open
and consuming memory.  And generating 'old request' warnings in the logs
(when the filestore is taking a long time to apply to the fs).

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>